decompile: upstream 8d47b585cf4b5e717aa06b2897d27c843fafa343 (2022-05-24) -- 158 files changed
Upstream commit: https://github.com/espressif/esp-thread-lib/commit/8d47b585cf4b5e717aa06b2897d27c843fafa343

Upstream subject: openthread: rebuild the lib with new toolchain

Archives decompiled: 2

+++ decompiled/libopenthread_br/esp_openthread_discovery.cpp/discovery_delegate_update.c
@@ -1,8 +1,8 @@
 /*
- * Last changed at upstream commit 8fdeda2b9b6e94761ab7fead714391e6bd27d486
- * https://github.com/espressif/esp-thread-lib/commit/8fdeda2b9b6e94761ab7fead714391e6bd27d486
- * Upstream date: 2021-09-09 20:40:32 +0800
- * Upstream subject: br: fix router solicitation handling(e82fe0d)
+ * Last changed at upstream commit 8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * https://github.com/espressif/esp-thread-lib/commit/8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * Upstream date: 2022-05-24 22:56:58 +0800
+ * Upstream subject: openthread: rebuild the lib with new toolchain
  * Source: libopenthread_br -> esp_openthread_discovery.cpp.o -> discovery_delegate_update
  *
  * (C) Espressif, Apache License 2.0.
@@ -22,7 +22,7 @@ void discovery_delegate_update(esp_openthread_mainloop_context_t *param_1)
   uVar1 = s_mdns_event_fd;
   if (-1 < (int)s_mdns_event_fd) {
     if ((int)s_mdns_event_fd < 0x40) {
-      peVar3 = param_1 + (s_mdns_event_fd >> 5) * 4;
+      peVar3 = param_1 + ((int)s_mdns_event_fd >> 5) * 4;
       uVar2 = 1 << (s_mdns_event_fd & 0x1f);
       *(uint *)peVar3 = *(uint *)peVar3 | uVar2;
       *(uint *)(peVar3 + 0x10) = uVar2 | *(uint *)(peVar3 + 0x10);
