+++ decompiled/libopenthread_port/esp_openthread_task_queue/esp_openthread_task_queue_update.c
@@ -1,8 +1,8 @@
 /*
- * Last changed at upstream commit 8fdeda2b9b6e94761ab7fead714391e6bd27d486
- * https://github.com/espressif/esp-thread-lib/commit/8fdeda2b9b6e94761ab7fead714391e6bd27d486
- * Upstream date: 2021-09-09 20:40:32 +0800
- * Upstream subject: br: fix router solicitation handling(e82fe0d)
+ * Last changed at upstream commit 8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * https://github.com/espressif/esp-thread-lib/commit/8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * Upstream date: 2022-05-24 22:56:58 +0800
+ * Upstream subject: openthread: rebuild the lib with new toolchain
  * Source: libopenthread_port -> esp_openthread_task_queue.o -> esp_openthread_task_queue_update
  *
  * (C) Espressif, Apache License 2.0.
@@ -19,7 +19,7 @@ void esp_openthread_task_queue_update(int param_1)
   uVar1 = s_task_queue_event_fd;
   if (-1 < (int)s_task_queue_event_fd) {
     if ((int)s_task_queue_event_fd < 0x40) {
-      puVar2 = (uint *)((s_task_queue_event_fd >> 5) * 4 + param_1);
+      puVar2 = (uint *)(((int)s_task_queue_event_fd >> 5) * 4 + param_1);
       *puVar2 = *puVar2 | 1 << (s_task_queue_event_fd & 0x1f);
     }
     if (*(int *)(param_1 + 0x18) < (int)uVar1) {
