+++ decompiled/libopenthread_port/esp_openthread_uart/esp_openthread_uart_update.c
@@ -1,8 +1,8 @@
 /*
- * Last changed at upstream commit 852eceaf38f6d6304f3b446a4a26f861389f83be
- * https://github.com/espressif/esp-thread-lib/commit/852eceaf38f6d6304f3b446a4a26f861389f83be
- * Upstream date: 2021-07-06 14:49:24 +0800
- * Upstream subject: openthread: make queue size and partition configurable
+ * Last changed at upstream commit 8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * https://github.com/espressif/esp-thread-lib/commit/8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * Upstream date: 2022-05-24 22:56:58 +0800
+ * Upstream subject: openthread: rebuild the lib with new toolchain
  * Source: libopenthread_port -> esp_openthread_uart.o -> esp_openthread_uart_update
  *
  * (C) Espressif, Apache License 2.0.
@@ -18,7 +18,7 @@ void esp_openthread_uart_update(int param_1)
   
   uVar1 = s_uart_fd;
   if (s_uart_fd < 0x40) {
-    puVar2 = (uint *)((s_uart_fd >> 5) * 4 + param_1);
+    puVar2 = (uint *)(((int)s_uart_fd >> 5) * 4 + param_1);
     *puVar2 = *puVar2 | 1 << (s_uart_fd & 0x1f);
   }
   if (*(int *)(param_1 + 0x18) < (int)uVar1) {
