+++ decompiled/libopenthread_port/esp_uart_spinel_interface.cpp/Process.c
@@ -1,8 +1,8 @@
 /*
- * Last changed at upstream commit 852eceaf38f6d6304f3b446a4a26f861389f83be
- * https://github.com/espressif/esp-thread-lib/commit/852eceaf38f6d6304f3b446a4a26f861389f83be
- * Upstream date: 2021-07-06 14:49:24 +0800
- * Upstream subject: openthread: make queue size and partition configurable
+ * Last changed at upstream commit 8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * https://github.com/espressif/esp-thread-lib/commit/8d47b585cf4b5e717aa06b2897d27c843fafa343
+ * Upstream date: 2022-05-24 22:56:58 +0800
+ * Upstream subject: openthread: rebuild the lib with new toolchain
  * Source: libopenthread_port -> esp_uart_spinel_interface.cpp.o -> Process
  *
  * (C) Espressif, Apache License 2.0.
@@ -20,7 +20,7 @@ esp::openthread::UartSpinelInterface::Process
   uint uVar1;
   
   uVar1 = *(uint *)(this + 0x4c);
-  if ((uVar1 < 0x40) && ((1 << (uVar1 & 0x1f) & *(uint *)(param_1 + (uVar1 >> 5) * 4)) != 0)) {
+  if ((uVar1 < 0x40) && ((1 << (uVar1 & 0x1f) & *(uint *)(param_1 + ((int)uVar1 >> 5) * 4)) != 0)) {
     TryReadAndDecode(this);
     return;
   }
